+++ src/views/Marketplace.vue
@@ -1,6 +1,11 @@
 <template>
   <div id="marketplace">
-    <ec-apps-market @click:app="goToApp" v-bind:tab="tab"/>
+    <ec-apps-market
+      @click:app="goToApp"
+      v-bind:tab="tab"
+      :additional-apps="mockedApps"
+      :next-tab="nextTab"
+    />
   </div>
 </template>