+++ src/views/js/Marketplace.js
@@ -1,8 +1,17 @@
-import EcAppsMarket from './../../components/EcAppsMarket.vue'
+import mockedApps from './../../lib/mocked-apps'
 import { genAppPath } from './../../router/'
+import EcAppsMarket from './../../components/EcAppsMarket.vue'
 
 export default {
   name: 'Marketplace',
+
+  computed: {
+    mockedApps: () => mockedApps,
+
+    nextTab () {
+      return this.$route.params.tab
+    }
+  },
 
   components: {
     EcAppsMarket